Init new version.  refactor of the package overall for better scalability of new functionalities to be added.
