+++ setup.py
@@ -5,16 +5,16 @@ except ImportError:
 from codecs import open
 
 setup(
-    name="statcode",
-    description="Like man pages, but for HTTP status codes",
-    version="v1.0.0",
+    name="httphelp",
+    description="Like man pages, but for HTTP status codes and headers (and more)",
+    version="v1.1.0",
     install_requires=["pyyaml", "urwid"],
     packages=["statcode"],
-    entry_points={"console_scripts": ["statcode = statcode.statcode:main"]},
+    entry_points={"console_scripts": ["httphelp = statcode.statcode:main"]},
     include_package_data=True,
     python_requires=">=3",
-    url="https://github.com/shobrook/statcode",
-    author="shobrook",
-    author_email="shobrookj@gmail.com",
+    url="https://github.com/Malex/statcode",
+    author="Malex",
+    author_email="malexprojects@gmail.com",
     license="MIT"
 )
